requirements.txt
torch==1.5.0
numpy==1.18.4
matplotlib==3.2.1
torchvision==0.6.0
pandas==1.0.4
Pillow==7.1.2
